all working need to check norm

+++ libft/ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: jheath <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/18 12:54:08 by jheath            #+#    #+#             */
-/*   Updated: 2018/06/02 21:09:25 by jheath           ###   ########.fr       */
+/*   Updated: 2018/06/04 10:06:50 by jheath           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlcat(char *dest, char *src, unsigned int size)
+unsigned int		ft_strlcat(char *dest, char *src, unsigned int size)
 {
 	size_t srcsize;
 	size_t destsize;
@@ -28,5 +28,5 @@ int		ft_strlcat(char *dest, char *src, unsigned int size)
 		ft_strncat(dest, src, size - destsize - 1);
 		dest[size - 1] = '\0';
 	}
-	return (destsize + srcsize);
+	return ((unsigned int)destsize + (unsigned int)srcsize);
 }
